+++ src/routes/tests/array-component/schema.ts
@@ -2,6 +2,6 @@ import { z } from 'zod';
 
 export const schema = z.object({
   sub: z.object({
-    tags: z.string().min(1).array().min(1)
+    tags: z.string().min(2).array().min(2)
   })
 });